riscv32toaout: fix warning

+++ Library/tools/riscv32toaout.c
@@ -50,7 +50,7 @@ static uint8_t *load_block(FILE *fp, off_t base, size_t len)
             fprintf(stderr, "Seek error.\n");
             exit(1);
         }
-        fprintf(stderr, "Loading %d bytes from %d\n", len, (int)base);
+        fprintf(stderr, "Loading %u bytes from %u\n", (unsigned)len, (unsigned)base);
         if (fread(m + 0x20, len, 1, fp) != 1) {
             fprintf(stderr, "Read error loading binary block.\n");
             exit(1);
